topic_analysis_service/requirements.txt
requests
requests_oauthlib
pymongo
nose
sklearn
pandas
feedparser
newspaper3k
praw
flask
scikit-multilearn
python-dateutil
majka
stop-words
